+++ src/CodeGenHero.DataService/Interfaces/IPageDataRequest.cs
@@ -4,7 +4,7 @@ namespace CodeGenHero.DataService
 {
 	public interface IPageDataRequest
 	{
-		List<IFilterCriterion> FilterCriteria { get; set; }
+		IList<IFilterCriterion> FilterCriteria { get; set; }
 		int Page { get; set; }
 		int PageSize { get; set; }
 		string Sort { get; set; }