+++ src/CodeGenHero.Logging/Models/LogMessageType.cs
@@ -24,6 +24,7 @@ namespace CodeGenHero.Logging
 		public virtual int Exception_Database => 102;
 		public virtual int Exception_Domain => 115;
 		public virtual int Exception_General => 103;
+		public virtual int Exception_Model => 117;
 		public virtual int Exception_Repository => 116;
 		public virtual int Exception_Synchronization => 110;
 		public virtual int Exception_Unhandled => 104;
@@ -36,8 +37,9 @@ namespace CodeGenHero.Logging
 		public virtual int Unauthorized => 401;
 		public virtual int Unknown => 100;
 		public virtual int Warn_Domain => 215;
-		public virtual int Warn_Repository => 216;
 		public virtual int Warn_Mobile => 206;
+		public virtual int Warn_Model => 217;
+		public virtual int Warn_Repository => 216;
 		public virtual int Warn_Synchronization => 210;
 		public virtual int Warn_Web => 207;
 		public virtual int Warn_WebApi => 205;